Compute graphlike distance given an error model (#171)
Performs a breadth first search over the graph defined by a graphlike error model, searching for the smallest set of errors that has no symptoms but does cause a logical error.

The actual graph being searched over is implicit. The nodes of the implicit graph have one or two detection events and a set of logical observable frame changes. The neighbors of a node are defined to be the states that can be reached by applying any error that touches the **single marked symptom that is allowed move**, producing tweaked symptoms and frame changes. The initial nodes that the search branches out from come from edges that actually cross an observable: the symptoms and frame changes of that edge define a starting node.

The reason for the restriction of neighbors to involve only one symptom at a time, until that symptom is gone, is that this still guarantees the optimal solution is reachable in the same number of steps but it can massively reduce the number of reachable/searched nodes. For example, consider a square patch with an observable across the middle. The starting state is an edge crossing the observable, with a symptom on the left and right. By forcing the left symptom to disappear before the right symptom can move, the reachable states are "right side symptom stays where it started, left side symptom anywhere" and then "left side symptom absorbed into boundary, right side symptom anywhere". The number of reachable states is roughly the area times 2. If both indices were allowed to move, then both symptoms could diffuse at the same time so all pairs of locations would be reachable so the number of reachable states would be proportional to the area squared.

The reason for starting the search from observable-crossing edges is because the solution must have one of those edges in it. In special cases it would be more efficient to do a "boundary to boundary" search, but unfortunately identifying whether or not the error model has the property that the boundary-to-boundary search will give the correct answer is highly non-trivial. For example, in a toric code there are no boundaries. Another example is that in a braiding based surface code some errors are cycles, not paths between boundaries. And another example is that in a surface code using lattice surgery and twists, whether or not a path between two boundaries is a logical error can depend on which specific route i

+++ src/stim/io/measure_record_reader.cc
@@ -399,7 +399,7 @@ bool MeasureRecordReaderFormatHits::start_record() {
                 "Bits per record is " + std::to_string(bits_per_record()) + " but got a hit value " +
                 std::to_string(value) + ".");
         }
-        buffer[value] ^= true;
+        buffer[(size_t)value] ^= true;
         is_first = false;
     }
     return true;
@@ -613,7 +613,7 @@ bool MeasureRecordReaderFormatDets::start_record() {
                 "Got '" + std::to_string(c) + std::to_string(number) + "' but expected num values of that type is " +
                 std::to_string(size) + ".");
         }
-        buffer[offset + number] ^= true;
+        buffer[(size_t)(offset + number)] ^= true;
     }
     return true;
 }
